Do not rely on Vue component mixins
